fix: scale debt shares exchange rate by 10^6
- [x] scale first debt shares minted by 10^6 to increase initial precision floor
- [x] change `total_debt0/1_shares` and `debt0/1_shares` type from `u64` to `u128`

+++ programs/omnipair/src/state/pair.rs
@@ -56,8 +56,8 @@ pub struct Pair {
     // Debt tracking (r_debt)
     pub total_debt0: u64,
     pub total_debt1: u64,
-    pub total_debt0_shares: u64,
-    pub total_debt1_shares: u64,
+    pub total_debt0_shares: u128,
+    pub total_debt1_shares: u128,
     
     // LP liquidity tracking
     pub total_supply: u64,
